:recycle: Divide serializer classes
