+++ coordinator/api/serializers/__init__.py
@@ -1,0 +1,6 @@
+from .release import ReleaseSerializer
+from .task import TaskSerializer
+from .task_service import TaskServiceSerializer
+from .study import StudySerializer
+from .release_note import ReleaseNoteSerializer
+from .event import EventSerializer
