+++ coordinator/api/serializers/study.py
@@ -1,0 +1,12 @@
+from rest_framework import serializers
+from coordinator.api.models import Study
+
+
+class StudySerializer(serializers.HyperlinkedModelSerializer):
+
+    version = serializers.CharField(source='latest_version', allow_blank=True)
+
+    class Meta:
+        model = Study
+        fields = ('kf_id', 'name', 'version', 'visible',
+                  'deleted', 'created_at')
